move map ploting from main.js to map-plot.js
